(PC-17817) chore(search): remarks
