+++ src/features/search/components/sections/Location/Location.native.test.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 
 import { Location } from 'features/search/components/sections/Location/Location'
 import { initialSearchState } from 'features/search/context/reducer'
-import { LocationType } from 'features/search/enums'
-import { RadioButtonLocation } from 'features/search/pages/modals/LocationModal/LocationModal'
+import { LocationType, RadioButtonLocation } from 'features/search/enums'
 import { GeoCoordinates } from 'libs/geolocation'
 import { render, fireEvent, superFlushWithAct } from 'tests/utils'
 import * as useModalAPI from 'ui/components/modals/useModal'
