+++ src/features/search/helpers/useShowResultsForCategory/useShowResultsForCategory.ts
@@ -6,9 +6,8 @@ import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useSearch } from 'features/search/context/SearchWrapper'
 import { OnPressCategory } from 'features/search/helpers/useSortedSearchCategories/useSortedSearchCategories'
+import { SearchView } from 'features/search/types'
 import { analytics } from 'libs/firebase/analytics'
-
-import { SearchView } from '../../types'
 
 export const useShowResultsForCategory = (): OnPressCategory => {
   const { searchState } = useSearch()
